feat: break out action creators
