+++ rrts/src/index.tsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { App } from './components/App';
-import { reducers } from './reducers/index';
+import { reducers } from './reducers';
 
 const store = createStore(reducers, applyMiddleware(thunk));
