+++ rrts/src/reducers/index.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { Todo } from '../actions/index';
+import { Todo } from '../actions';
 import { todosReducer } from './todos';
 
 export interface StoreState {
